+++ package.json
@@ -24,7 +24,6 @@
     "plotly.js": "^3.0.1",
     "react": "^19.0.0",
     "react-bootstrap": "^2.10.9",
-    "react-dom": "^19.0.0",
     "react-plotly.js": "^2.6.0",
     "react-bootstrap-icons": "^1.11.5"
   },
